dkroutingtool/requirements.txt
asn1crypto==1.4.0
backcall==0.2.0
boto3==1.16.17
botocore==1.19.17
branca==0.4.1
certifi==2020.6.20
#cffi==1.12.2
chardet==3.0.4
cryptography==3.2
decorator==4.4.0
docutils==0.14
et-xmlfile==1.0.1
flask==2.1.3
folium==0.9.1
geojson==2.4.1
idna==2.8
ipython==7.5.0
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.13.3
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
numpy==1.16.4
openpyxl==2.6.2
ortools==7.8.7959
pandas==1.1.1
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.9
protobuf==3.13.0
ptyprocess==0.6.0
pybind11==2.3.0
pycosat==0.6.3
pycparser==2.19
Pygments==2.4.2
pyOpenSSL==19.0.0
PySocks==1.6.8
pytest==7.1.2
python-bond==1.4
python-dateutil==2.8.0
pytz==2019.1
requests==2.23.0
s3transfer==0.3.3
scipy==1.3.0
six==1.13.0
traitlets==4.3.2
ujson==1.35
urllib3==1.25.8
wcwidth==0.1.7
xlrd==1.2.0
notebook==6.1.5
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.23.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
oauth2client==3.0.0
googleapis-common-protos==1.52.0
ruamel.yaml==0.16.12